refactor: merge remote and process recorders for node
These two recorders are very similar and can be merged to reduce code 
duplication.
