--- components/recorder-cli/remote/.env
@@ -1,1 +1,0 @@
-node
