+++ lib/node/recorder.mjs
@@ -3,9 +3,17 @@ import "./error.mjs";
 import { configuration } from "./configuration.mjs";
 export * from "./loader-esm.mjs";
 
+const map = {
+  __proto__: null,
+  remote: "node",
+  process: "node",
+  mocha: "mocha",
+  jest: "jest",
+};
+
 // Use top-level await to wait for ./configuration.mjs to update env variables.
 const { record } = await import(
-  `../../dist/bundles/recorder-${configuration.recorder}.mjs`
+  `../../dist/bundles/recorder-${map[configuration.recorder]}.mjs`
 );
 
 record(configuration);
